Moving GenericTableEntity tests into their own file.
